generative-ai/3-practice/3-4-customer-support-rag-with-eval/pyproject.toml
[project]
name = "3-4-customer-support-rag-with-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.11"
dependencies = [
    "langchain-community>=0.3.10",
    "langchain>=0.3.10",
    "streamlit>=1.40.2",
    "pandas>=2.2.3",
    "langchain-openai>=0.2.11",
    "langchain-chroma>=0.1.4",
    "onnxruntime==1.19.0",
]

[dependency-groups]
dev = [
    "remote-pdb>=2.1.0",
]